refactor(webos-common): split writeUtils into pseudoWriter and generateWriter
Separate writeUtils.js into two focused modules:
 - pseudoWriter.js: writePseudoResources (localize mode pseudo handling)
 - generateWriter.js: filterGenResources, writeGenResources (generate mode)
