+++ packages/ilib-loctool-webos-c/test/integrationTest/xliffs/ko-KR.xliff
@@ -59,6 +59,12 @@
           <target>재시도(universal)</target>
         </segment>
       </unit>
+      <unit id="3">
+        <segment>
+          <source>Time Settings</source>
+          <target>시간 설정(universal)</target>
+        </segment>
+      </unit>
     </group>
   </file>
 </xliff>
