+++ packages/ilib-loctool-webos-common/generateWriter.js
@@ -1,5 +1,5 @@
 /*
- * writeUtils.js - Utilities for write() mode resource handling
+ * generateWriter.js - Resource filtering and writing for generate mode
  *
  * Copyright (c) 2026 JEDLSoft
  *
@@ -59,42 +59,6 @@ function filterGenResources(resources, projectName, selfDatatype, options) {
     });
 
     return Object.keys(best).map(function(k) { return best[k].resource; });
-}
-
-/**
- * Collect, filter, and write pseudo resources to their resource files in localize mode.
- *
- * Collects pseudo resources from the pseudo bundle, filters by datatype and the
- * disablePseudo setting, then writes each qualifying resource to its resource file.
- *
- * @param {Object} pseudo - plugin's pseudo bundle (this.pseudo)
- * @param {Object} settings - project settings (project.settings)
- * @param {string} settingsKey - key into settings for this plugin type (usually this.type;
- *   for Dart use this.type.replace("x-", ""))
- * @param {string} datatype - plugin's own datatype for filtering (this.datatype)
- * @param {Object} resFileType - resource file type from project.getResourceFileType()
- * @param {string} sourceLocale - project source locale
- * @param {string} deviceType - device type for target resolution
- * @param {Function} [resPathFn] - optional callback(res) returning localized resource path.
- *   Required for plugins where getResourceFile() takes a path argument (e.g. Dart).
- */
-function writePseudoResources(pseudo, settings, settingsKey, datatype, resFileType, sourceLocale, deviceType, resPathFn) {
-    var pseudoResources = [];
-    if (settings[settingsKey] === undefined ||
-        (settings[settingsKey] && !(settings[settingsKey].disablePseudo === true))) {
-        pseudoResources = pseudo.getAll().filter(function(res) {
-            return res.datatype === datatype;
-        });
-    }
-
-    pseudoResources.forEach(function(res) {
-        if (res.getTargetLocale() !== sourceLocale &&
-            res.getSource() !== pluginUtils.getTarget(res, deviceType)) {
-            var resPath = resPathFn ? resPathFn(res) : undefined;
-            res.setTarget(pluginUtils.getTarget(res, deviceType));
-            resFileType.getResourceFile(res.getTargetLocale(), resPath).addResource(res);
-        }
-    });
 }
 
 /**
@@ -184,6 +148,5 @@ function writeGenResources(project, translationLocales, genresources, params) {
 
 module.exports = {
     filterGenResources: filterGenResources,
-    writePseudoResources: writePseudoResources,
     writeGenResources: writeGenResources
 };
