+++ packages/ilib-loctool-webos-dart/DartFileType.js
@@ -24,12 +24,12 @@ var Locale = require("ilib/lib/Locale.js");
 var DartFile = require("./DartFile.js");
 var JsonResourceFileType = require("ilib-loctool-webos-json-resource");
 var Utils = require("loctool/lib/utils.js")
-var { utils: pluginUtils, translationResolver, writeUtils } = require("ilib-loctool-webos-common");
+var { utils: pluginUtils, translationResolver, pseudoWriter, generateWriter } = require("ilib-loctool-webos-common");
 var buildResolver = translationResolver.buildResolver;
 var resolveTranslation = translationResolver.resolveTranslation;
-var filterGenResources = writeUtils.filterGenResources;
-var writePseudoResources = writeUtils.writePseudoResources;
-var writeGenResources = writeUtils.writeGenResources;
+var filterGenResources = generateWriter.filterGenResources;
+var writePseudoResources = pseudoWriter.writePseudoResources;
+var writeGenResources = generateWriter.writeGenResources;
 
 var DartFileType = function(project) {
     this.type = "x-dart";
